[py] move DFtoVW to its own file
- split test file also
